fix use of forward declared types
